Converted all floats to doubles
Prepared simulation environment

+++ Packages/com.philsa.ecs-physics/Code/Systems/PhysicsSystem.cs
@@ -16,10 +16,10 @@ namespace PhysicsEngine
     [UpdateAfter(typeof(UnityEngine.Experimental.PlayerLoop.FixedUpdate))]
     public class PhysicsSystem : JobComponentSystem
     {
-        public float DeltaTime;
+        public double DeltaTime;
 
         public PhysicsSettings Settings;
-        
+
         public NativeArray<CollisionPair> SphereSphereCollisionPairsArray;
         public NativeArray<CollisionPair> SphereBoxCollisionPairsArray;
         public NativeArray<CollisionPair> BoxBoxCollisionPairsArray;
